updated JavaScript
Created a function called validdateSyntax, get go get my input on HTML which is "petInput" and then access the current value entered on my input field. And I had to ensure that the input string starts with and is followed by four digits which are (`\d{4}`) then followed by one or more alphanumeric characters (`[A-Za-zO-9]+`).